requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
pillow
transformers
tqdm

